backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
psycopg
psycopg-binary
pandas
